requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
pinecone
redis
yt-dlp
openai-whisper
youtube-transcript-api
numpy
scikit-learn